refactor: move shared integration contracts to src/shared/types/integrations.ts
No behavior change: build clean, 2619/2619 tests pass, lint clean.
